+++ tomcat/src/test/java/org/apache/coyote/http11/request/HttpRequestStartLineTest.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.request;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -8,8 +8,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 @SuppressWarnings("NonAsciiCharacters")
-@DisplayName("HttpStartLine 테스트")
-class HttpStartLineTest {
+@DisplayName("HttpRequestStartLine 테스트")
+class HttpRequestStartLineTest {
 
     @Test
     void HTTP_요청_메세지_시작라인의_정보를_저장한다() {
@@ -17,13 +17,13 @@ class HttpStartLineTest {
         final String startLine = "GET /index.html HTTP/1.1";
 
         // when
-        final HttpStartLine httpStartLine = HttpStartLine.from(startLine);
+        final HttpRequestStartLine httpRequestStartLine = HttpRequestStartLine.from(startLine);
 
         // then
         assertSoftly(softAssertions -> {
-            assertThat(httpStartLine.getHttpRequestMethod()).isEqualTo(HttpRequestMethod.GET);
-            assertThat(httpStartLine.getRequestURI()).isEqualTo("/index.html");
-            assertThat(httpStartLine.getHttpVersion()).isEqualTo("HTTP/1.1");
+            assertThat(httpRequestStartLine.getHttpRequestMethod()).isEqualTo(HttpRequestMethod.GET);
+            assertThat(httpRequestStartLine.getRequestURI()).isEqualTo("/index.html");
+            assertThat(httpRequestStartLine.getHttpVersion()).isEqualTo("HTTP/1.1");
         });
     }
 
@@ -33,7 +33,7 @@ class HttpStartLineTest {
         final String startLine = "GETHTTP/1.1";
 
         // when & then
-        assertThatThrownBy(() -> HttpStartLine.from(startLine))
+        assertThatThrownBy(() -> HttpRequestStartLine.from(startLine))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessage("시작 라인의 토큰은 3개여야 합니다.");
     }
